feat: finish project

+++ components/LikeButton.tsx
@@ -1,51 +1,43 @@
-'use client';
+"use client";
+import React, { useEffect, useState } from "react";
+import { useSessionContext } from "@supabase/auth-helpers-react";
+import { useRouter } from "next/navigation";
+import { AiFillHeart, AiOutlineHeart } from "react-icons/ai";
+import toast from "react-hot-toast";
 
-import React, { useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
-
-import { useSessionContext } from '@supabase/auth-helpers-react';
-import useAuthModal from '@/hooks/useAuthModal';
-import { useUser } from '@/hooks/useUser';
-import { AiFillHeart, AiOutlineHeart } from 'react-icons/ai';
-import { toast } from 'react-hot-toast';
+import useAuthModal from "@/hooks/useAuthModal";
+import { useUser } from "@/hooks/useUser";
 
 interface LikeButtonProps {
   songId: string;
 }
 
-const LikeButton: React.FC<LikeButtonProps> = ({
-                                                 songId
-                                               }) => {
+const LikeButton: React.FC<LikeButtonProps> = ({ songId }) => {
   const router = useRouter();
-  const {supabaseClient} = useSessionContext();
+  const { supabaseClient } = useSessionContext();
 
   const authModal = useAuthModal();
-  const {user} = useUser();
+  const { user } = useUser();
 
-  const [isLiked, setIsLiked] = useState<boolean>(false);
-
+  const [isLiked, setIsLiked] = useState(false);
   useEffect(() => {
     if (!user?.id) {
       return;
     }
 
     const fetchData = async () => {
-      const {
-        data,
-        error
-      } = await supabaseClient
-        .from('liked_songs')
-        .select('*')
-        .eq('user_id', user.id)
-        .eq('song_id', songId)
+      const { data, error } = await supabaseClient
+        .from("liked_songs")
+        .select("*")
+        .eq("user_id", user.id)
+        .eq("song_id", songId)
         .single();
 
       if (!error && data) {
         setIsLiked(true);
       }
     };
-
-    fetchData().then();
+    fetchData();
   }, [songId, supabaseClient, user?.id]);
 
   const Icon = isLiked ? AiFillHeart : AiOutlineHeart;
@@ -54,13 +46,8 @@ const LikeButton: React.FC<LikeButtonProps> = ({
     if (!user) {
       return authModal.onOpen();
     }
-
     if (isLiked) {
-      const {error} = await supabaseClient
-        .from('liked_songs')
-        .delete()
-        .eq('user_id', user.id)
-        .eq('song_id', songId);
+      const { error } = await supabaseClient.from("liked_songs").delete().eq("user_id", user.id).eq("song_id", songId);
 
       if (error) {
         toast.error(error.message);
@@ -68,33 +55,19 @@ const LikeButton: React.FC<LikeButtonProps> = ({
         setIsLiked(false);
       }
     } else {
-      const {error} = await supabaseClient
-        .from('liked_songs')
-        .insert({
-          user_id: user.id,
-          song_id: songId
-        });
-
+      const { error } = await supabaseClient.from("liked_songs").insert({ song_id: songId, user_id: user.id });
       if (error) {
         toast.error(error.message);
       } else {
         setIsLiked(true);
-        toast.success('Liked!');
+        toast.success("Liked!");
       }
     }
-
     router.refresh();
   };
-
   return (
-    <button
-      className="hover:opacity-75 transition"
-      onClick={handleLike}
-    >
-      <Icon
-        color={isLiked ? '#22c55e' : 'white'}
-        size={25}
-      />
+    <button onClick={handleLike} className="hover:opacity-75 transition">
+      <Icon color={isLiked ? "#22c55e" : "#fff"} size={25} />
     </button>
   );
 };
